Update to the article on using PSRAM on Zephyr
This work comes to update necessary details to use PSRAM with more recent
Zephyr versions. It also updates the log results to be consistent with
those observed when running the samples with more recent Zephyr versions.
This work also adds a session explaining how to place Zephyr's task stacks
in external RAM and the associated risks.

Signed-off-by: Marcio Ribeiro <marcio.ribeiro@espressif.com>